Move extend methods from AstSchemaBuilder to AstSchemaExtender
